I decided not to bind the token to the app. object.
Since it's needed for both the authorization code and password grant
flow (and we'll see for the client credentials flow), there really
isn't a _need_ to attach the token to the app. object since that
requires you to remember which token you just used the app. to grab
when it's much easier to just return the token and then have the user
store it, etc., whatever with it rather than remembering, "Did I save
the authorization code token and should I grab it from the app. object
_now_ before updating that object with the password grant token and
_then_ grabbing it from the object to do something with it?"