On second thought, it might be easier with the token as part of the app.
It's easier for me since just the app. has to be passed in and,
hopefully, easier for the user since they don't have to keep track of
the token separately, that way.

I suppose they need to save it for local cache but they just load a
single variable up and then can pass it in, anywhere.

Since it's set to return the app. – after it updates it –, hopefully
that'll keep things simplified and streamlined for the user.

It might not be pleasant if they're trying to use multiple grants in a
concurrent environment but hopefully that won't be the case for the
same app. in the same instance for a program.