And to make the calls for filters.
I need to check if any of the previous calls, essentially, used a
masto object to derive how to handle the HTTP call but, I figure, why
limit the user to not make full use of the objects they're handling?

This way, say you grab a filter; then you decide to update an element
of it. Rather than handling all the parts of it, you just update the
part in question and pass the object along.

On that note, I should probably make it editable, then…; let's start
here and keep the ones you don't need to mess with inaccessible.