I think I'm going to do this a different way but I want to commit the original idea before tweaking it.
Basically, I realized that, if I passed an empty list, the section
that creates an array of the query params would error out since there
is no first index of a null string.

So this checks before trying to get rid of the initial ampersand.

Also, – while the main list gets filtered for any false values – the
array's list never does so add that, as well, so I don't have to check
parameters for if they're false.