Update filters…
I realized that "expires_in" takes an amount of _seconds_, rather than
a date…, so I had to update that section, as well.

While I don't know that you would necessarily ever be able to
organically end up at this spot, – since I want the user to be able to
pass a filter record object to `masto-filter-create`, if so desired –
the user could store a date object in the `expiresAt` spot.

The function will then just do the math for them and calculate the
number of second between now and when the date it should expire at is.

It will also handle a time object or just the number of
seconds. This'll provide full support to any kind of reasonable object
types that might be used to represent this data.