Well, this solves, somewhat, my wondering if it was worth moving different elements of `entities` into different files.
While it doesn't solve the sheer size issue, it does solve the
compartmentalizing for the end user.

I rather prefer this than the end user importing every single object
at once.

The relevant stuff remains in the finally used module and the
interdependency of the different records isn't disrupted (or have to
be planned around as the elements move to different files).