Turns out there was a name conflict so renaming the record internal function.
This makes me wonder if I should make a different name for functions
that grab elements of a record and functions that grab from the API.

I suppose they're technically doing the same thing, even if the
background processes are different, but I could see someone getting
confused by `masto-status-reblog` vs. `masto-status-reblog-status`.

But that would require rewriting the majority of the functions…

I suppose, since they're just the API calls (and, therefore, defined
and exported in only one file, each), that wouldn't be as terrible as
it could be but…ugh, I'll think on it, later.