A script to generate documentation. It's not the best and I feel like I should have better tools than this for a Scheme…but it's sure as Hell better than nothing.
This way I can provide commentary for modules and documentation in
docstrings for the functions and this script will be able to grab all
of that for me without me having to update more than one place.

And, while I'd prefer being able to better capture the nature of, say,
arguments (like not just which are keyword arguments but if those
keyword arguments have defaults), that'll get caught by my
documentation in the docstring, anyway (since I don't believe, for
example, Geiser ends up specifying whether the arguments have
defaults, either), and so it's still capturing all the absolutely
crucial elements, even if it's a bit more work for my end.

The big thing is just not having to update more than one place; that's
what I wanted the most, in the end.