And a bit more; if there aren't any modules which have functions or records that appear in the documentation string, just pass back the documentation.
If there are, find the module that has the function we're currently
documenting let's go through every function/record in our current
module and remove those values from the other modules in the list.

We're doing this since, if the function or record we want to reference
resides in the current documentation page, we'll want to reference to
this page, rather than a different page (so, if the `statuses`
documentation page references `<mastodon-account>`s, we want to
reference the documentation for that that's on this page, rather than
its documentation on the `accounts` page).

It's…ugly as Hell but I think that's as simplified as I can write that
code.