As we're going through and removing functions/exports from the other modules, we can use the function/export we're checking the other modules for to replace all(, if any,) instances of it in the documentation with a reference to the current module.
I hate introducing a destructive function but I don't know how I'd
easily keep track of changes (and easily save them to a variable I can
reference, later) for `documentation` without doing something like
saving it in the list we're `fold`ing into.

And I'm not going to muddy things up, that much, for functional purity
in a script.