There may be a cleaner or more straightforward way to do this but, again, I'm not going to bother for a script.
Take the list we now have in `remainingModules` and remove all
duplicate functions/records.

We still need to keep track of which module they belong to so we're
going to, basically, just end up with a simple alist (with a lot of
duplicate keys but unique values).

We don't have to look anything up since we'll have to cycle through
them because we have to cover all of them so this doesn't matter.