Add documentation for scheduled statuses.
I…dunno how to handle the `-set!` functions; I could do it with a
negative lookaround but it doesn't seem that Guile's regexp. has that
option.

There really isn't a way to filter for it since I need to process the
shorter function name, as well. I can't even say something like
"<function name>(.*[^`][^\\]][^\\(])" because, if the function's at the end of
a line, that's not going to get counted and not get replaced.

So…I dunno; I'll just manually update, I guess…